# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Privy Stream API
  version: 1.0.0
  description: Внешний REST API Gateway для аудио-стримингового сервиса
servers:
  - url: http://localhost:8080
paths:
...
  /catalog/tracks/{track_uuid}:
    get:
      summary: Получить путь до файла трека и его длительность
      operationId: getTrackById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackUuid'
      responses:
        '200':
          description: Путь до аудиофайла и длительность
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackPath'
        '401':
          description: Missing or invalid authorization header
        '404':
          description: Track not found
    delete:
      summary: Удалить трек
      description: >
        Удаляет трек из каталога, его позицию в трек-листе альбома и вхождения
        в плейлисты, а также файл из хранилища узла. Файл обложки не удаляется:
        он принадлежит альбому и его остальным трекам.
      operationId: deleteTrack
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackUuid'
      responses:
        '204':
          description: Трек удалён
        '400':
          description: Invalid track uuid
        '401':
          description: Missing or invalid authorization header
        '404':
          description: Track not found
  /catalog/tracks/{track_uuid}/exists:
    get:
      summary: Проверить существование трека
      operationId: trackExists
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackUuid'
      responses:
        '200':
          description: Результат проверки
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
        '401':
          description: Missing or invalid authorization header
  /catalog/tracks:
    post:
      summary: Добавить трек
      operationId: addTrack
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTrackRequest'
      responses:
        '201':
          description: Трек добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackDetails'
        '400':
          description: Invalid request data
        '401':
          description: Missing or invalid authorization header
        '409':
          description: Track already exists
  /catalog/tracks/{track_uuid}/file:
    post:
      summary: Загрузить аудиофайл трека
      operationId: addTrackFile
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackUuid'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Файл загружен
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  size:
                    type: integer
        '400':
          description: Invalid request body
        '401':
          description: Missing or invalid authorization header
        '404':
          description: Track not found
        '500':
          description: Internal server error
  /catalog/tracks/{track_uuid}/cover:
    post:
      summary: Загрузить обложку трека (становится обложкой его альбома)
      description: >
        Файл один, ссылку на него получают и трек, и его альбом со всеми треками:
        обложка трека всегда совпадает с обложкой альбома. Имя файла на сервере
        генерируется заново, у клиента берётся только расширение (jpg, jpeg, png, webp, gif).
        Итоговый путь возвращается в ответе и читается через
        GET /catalog/albums/{album_uuid} (поле cover_path).
      operationId: addTrackCover
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackUuid'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Обложка загружена
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  size:
                    type: integer
        '400':
          description: Invalid request body
        '401':
          description: Missing or invalid authorization header
        '404':
          description: Track not found
        '500':
          description: Internal server error
...
  /catalog/artists/{artist_uuid}:
    get:
      summary: Получить артиста по ID
      operationId: getArtistById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArtistUuid'
      responses:
        '200':
          description: Артист
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '401':
          description: Missing or invalid authorization header
        '404':
          description: Artist not found
  /catalog/artists/{artist_uuid}/albums:
    get:
      summary: Получить альбомы артиста
      operationId: getArtistAlbums
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArtistUuid'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список альбомов артиста
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LightAlbum'
        '401':
          description: Missing or invalid authorization header
...
  /catalog/artists/{artist_uuid}/tracks:
    get:
      summary: Получить треки артиста
      operationId: getArtistTracks
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArtistUuid'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Список треков артиста
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LightTrack'
        '401':
          description: Missing or invalid authorization header
  /catalog/artists:
    post:
      summary: Добавить артиста
      operationId: addArtist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddArtistRequest'
      responses:
        '201':
          description: Артист добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Invalid request data
        '401':
          description: Missing or invalid authorization header
        '409':
          description: Artist already exists
  /catalog/albums/{album_uuid}:
    get:
      summary: Получить альбом по ID
      operationId: getAlbumById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumUuid'
      responses:
        '200':
          description: Альбом
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '401':
          description: Missing or invalid authorization header
        '404':
          description: Album not found
    delete:
      summary: Удалить альбом
      description: >
        Удаляет альбом вместе со всеми его треками: позиции в трек-листе,
        вхождения треков в плейлисты и файлы из хранилища (файлы треков и
        обложка). Артист остаётся. Отменить нельзя.
      operationId: deleteAlbum
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumUuid'
      responses:
        '204':
          description: Альбом удалён
        '400':
          description: Invalid album uuid
        '401':
          description: Missing or invalid authorization header
        '404':
          description: Album not found
  /catalog/albums/{album_uuid}/tracks:
    get:
      summary: Получить треки альбома
      operationId: getAlbumTracks
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumUuid'
      responses:
        '200':
          description: Список треков альбома
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LightTrack'
        '401':
          description: Missing or invalid authorization header

    post:
      summary: Привязать треки к альбому (с указанием позиции в трек-листе)
      operationId: addTracksToAlbum
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumUuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTracksToAlbumRequest'
      responses:
        '200':
          description: Треки привязаны к альбому
        '400':
          description: Invalid request data
        '401':
          description: Missing or invalid authorization header
        '404':
          description: Album or track not found
        '409':
          description: Track already in album, or position already taken
  /catalog/albums/{album_uuid}/cover:
    get:
      summary: Получить файл обложки альбома
      description: >
        Отдаёт содержимое файла по cover_path альбома (см. GET /catalog/albums/{album_uuid}).
        Как GET /stream/{id}: <img> не отправляет Authorization, поэтому клиент
        скачивает картинку авторизованным запросом и показывает из памяти.
      operationId: getAlbumCover
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumUuid'
      responses:
        '200':
          description: Содержимое файла обложки
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid album uuid
        '401':
          description: Missing or invalid authorization header
        '404':
          description: Album not found, or album has no cover
        '500':
          description: Internal server error
    post:
      summary: Загрузить обложку альбома
      description: >
        Путь обложки записывается альбому и всем его трекам (обложка альбома
        приоритетна). Имя файла на сервере генерируется заново, у клиента
        берётся только расширение (jpg, jpeg, png, webp, gif).
        Итоговый путь возвращается в ответе и читается через
        GET /catalog/albums/{album_uuid} (поле cover_path).
      operationId: addAlbumCover
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumUuid'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Обложка загружена
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  size:
                    type: integer
        '400':
          description: Invalid request body
        '401':
          description: Missing or invalid authorization header
        '404':
          description: Album not found
        '500':
          description: Internal server error
  /catalog/albums:
    post:
      summary: Добавить альбом
      operationId: addAlbum
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAlbumRequest'
      responses:
        '201':
          description: Альбом добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Invalid request data
        '401':
          description: Missing or invalid authorization header
        '409':
          description: Album already exists
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        default: 0
    TrackUuid:
      name: track_uuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ArtistUuid:
      name: artist_uuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AlbumUuid:
      name: album_uuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    LightTrack:
      type: object
      properties:
        track_uuid:
          type: string
          format: uuid
        track_name:
          type: string
          example: "Убийца карася"
        explicit:
          type: boolean
        duration_ms:
          type: integer
    TrackPath:
      type: object
      properties:
        path:
          type: string
        duration_ms:
          type: integer
        album_uuid:
          type: string
          format: uuid
        cover_path:
          type: string
          description: Путь файла обложки в хранилище; всегда совпадает с cover_path альбома, пустая строка — обложки нет
    TrackDetails:
      type: object
      description: Плоская сущность трека (без артиста/альбома, только связи по id) — то, что реально хранится в таблице tracks
      properties:
        track_uuid:
          type: string
          format: uuid
        track_name:
          type: string
          example: "Убийца карася"
        artist_uuid:
          type: string
          format: uuid
        album_uuid:
          type: string
          format: uuid
        explicit:
          type: boolean
        path:
          type: string
        duration_ms:
          type: integer
    AddTrackRequest:
      type: object
      required:
        - track_name
        - artist_uuid
        - album_uuid
      properties:
        track_name:
          type: string
          example: "Убийца карася"
        artist_uuid:
          type: string
          format: uuid
        album_uuid:
          type: string
          format: uuid
        explicit:
          type: boolean
        path:
          type: string
          description: |
            Необязательная подсказка расширения файла (например, `track.flac`).
            Итоговый путь генерирует сервер: `{album_uuid}/{uuid}{ext}`. Из присланного значения
            берётся только расширение из списка .mp3, .flac, .ogg, .opus, .m4a, .aac, .wav,
            остальное игнорируется. Фактический путь возвращается в ответе (`TrackDetails.path`).
        duration_ms:
          type: integer
    Artist:
      type: object
      properties:
        artist_uuid:
          type: string
          format: uuid
        artist_name:
          type: string
          example: "PlayBoyCalich"
    AddArtistRequest:
      type: object
      required:
        - artist_name
      properties:
        artist_name:
          type: string
          example: "PlayBoyCalich"
    LightAlbum:
      type: object
      properties:
        album_uuid:
          type: string
          format: uuid
        album_name:
          type: string
          example: "Убийца карася"
        created_at:
          type: string
          format: date-time
    Album:
      type: object
      description: Плоская сущность из таблицы albums — треки и данные артиста запрашиваются отдельно (GetAlbumTracks / GetArtistByID)
      properties:
        album_uuid:
          type: string
          format: uuid
        artist_uuid:
          type: string
          format: uuid
        album_name:
          type: string
          example: "Убийца карася"
        created_at:
          type: string
          format: date-time
        cover_path:
          type: string
          description: |
            Путь файла обложки в хранилище, пустая строка — обложки нет.
            Заполняется через POST /catalog/albums/{album_uuid}/cover или
            POST /catalog/tracks/{track_uuid}/cover: файл один, и у альбома,
            и у всех его треков путь совпадает.
    AddAlbumRequest:
      type: object
      required:
        - artist_uuid
        - album_name
      properties:
        artist_uuid:
          type: string
          format: uuid
        album_name:
          type: string
          example: "Убийца карася"
    AlbumTrackInput:
      type: object
      required:
        - track_uuid
        - position
      properties:
        track_uuid:
          type: string
          format: uuid
        position:
          type: integer
          minimum: 1
    AddTracksToAlbumRequest:
      type: object
      required:
        - tracks
      properties:
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/AlbumTrackInput'